tsbartProbit: Partially Smooth BART for binary responses

This function implements the probit formulation of Partially Smooth BART method for a binary response, smoothing over the covariate tgt, with unsmoothed covariates x. Predicts function values at a set of new time points and covariates.

tsbartProbit(y, yobs, tgt, tpred, x, xpred, xinfo_list,
           nburn, nsim, offset, m, ecross, base_tree, power_tree,
           treef_name_, save_trees, silent_mode)

Arguments

y
Length n vector with continuous latent y values. If unknown, set to 1.96 if yobs=1 and -1.96 if yobs=0.

yobs
Length n vector with observed binary 1/0 responses. Required,

tgt
Length n targeted covariate over which to smooth.

tpred
Length npred out-of-sample targeted covariate.

x
A (n x p) data frame or matrix containing the covariates which are not to be smoothed over. Observations must be in columns. If data frame is set up in usual observations (rows) and covariates (columns) form, input t(x).

xpred
A (npred x p) data frame or matrix containing out of sample covariates. Observations must be in columns. If data frame is set up in usual observations (rows) and covariates (columns) form, input t(x).

xinfo_list
A list with length matching number of covariates in x. Each element contains a vector of cutpoints. Can be generated using the helper function makeCutPoints(). Required,

nburn
Number of burn-in MCMC iterations.

nsim
Number of MCMC iterations to save after burn-in.

offset
Double, holds offset for probit. See examples for setting offset.

ntree
Number of trees. Defaults to 200.

nu
Degrees of freedom in the chisq prior on sigma^2. Defaults to 3.

ecross
Double number of expected times f(x,t) crosses the mean response over time, alpha(t). Required, We suggest a default value of 1.

base_tree
Base for tree prior. Defaults to 0.95.

power_tree
Power for tree prior. Defaults to 2.0.

treef_name_
Character vector containing file name/path for tree output. Must be populated, even if save_trees is set to FALSE. (In this case, parameter is ignored.)

save_trees
Boolean for writing trees to a .txt file.

silent_mode
Boolean for writing progress report to console.

Value

A list containing the following items:

mcmcdraws

A (nsim x n) matrix containing the in-sample MCMC draws for the Partially Smooth BART fit.

mcmcdraws_oos

A (nsim x npred) matrix containing the out-of-sample MCMC draws for the Partially Smooth BART fit.

eta

A vector containing the MCMC draws for eta .

gamma

A vector containing the MCMC draws for gamma .

postpred

A (nsim x n) matrix of posterior predictive draws for in-sample data.

postpred_oos

A (nsim x npred) matrix of posterior predictive draws for out-of-sample data.

Examples

# NOT RUN {
Function call example is as follows.

# Calibrate BART's error variance a la CGM 2010 (method 2)
df = data.frame(xx, y)
lmf = lm(y~.,df)
sighat = sigma(lmf) #sd(y) #summary(lmf)$sigma

# Hyperparameters
nu = 3
sigq = .9
qchi = qchisq(1.0-sigq,nu)
lambda = (sighat*sighat*qchi)/nu

# Decision rule cutpoints for each predictor.
cutpoints = makeCutpoints(xx)

# Calculate offset
phat = mean(yobs)
offset = qnorm(phat)

# Fit model
fit = tsbartProbit(y = y,
                   y = yobs,
                   tgt = ti,
                   tpred = ti_pred,
                   x = t(xx),         # observations must be in columns
                   xpred = t(xpred), # observations must be in columns
                   xinfo_list = cutpoints,
                   nburn=1000, nsim=10000, ntree=200, nu=nu, ecross=1, base_tree=.9, power_tree=2,
                   treef_name_ = "mytrees.csv",
                   save_trees=F, scale_y=T, silent_mode=F)

# }