tsbcfProbit: tsbart for causal inference for binary responses

This function implements the tsbart method for a binary response in the causal inference setting, smoothing over the covariate tgt, with unsmoothed covariates x and treatment indicator z. Predicts function values f(x,t,z) at a set of new time points and covariates.

Augmented model formulation is: y(t) = alpha(t) + eta_m * m(x,t) + eta_a * a(x,t)*z + epsilon

where propensity scores should be included as part of each x matrix of covariates. The output "pmean" gives y-hat(t). The output "pmean_trt" gives (eta_a * a(x,t)*z). Probability of success given x and treatment is pnorm(y(t)).

tsbcfProbit(y, yobs, x, x_trt, tgt, z,
           xpred, xpred_trt, tpred, zpred,
           xinfo_list, xinfo_list_trt,
           nburn, nsim, ntree,
           lambda, nu, ecross, ecross_trt,
           treef_name_, save_trees, silent_mode)

Arguments

y
A vector with continuous response values. Required,

y_obs
A vector with observed binary 1/0 responses. Required,

z
A vector of 0/1 indicators for each observation. 1 = treatment group, 0 = control group. Required,

x
A data frame or matrix containing the covariates for the prognostic effect which are not to be smoothed over. Observations must be in columns. If data frame is set up in usual observations (rows) and covariates (columns) form, input t(x). Required,

x_trt
A data frame or matrix containing the covariates for the treatment effect which are not to be smoothed over. Observations must be in columns. If data frame is set up in usual observations (rows) and covariates (columns) form, input t(x). Required,

tgt
A vector with discrete time points. Required,

xpred
A data frame or matrix containing out of sample covariates. Observations must be in columns. If data frame is set up in usual observations (rows) and covariates (columns) form, input t(x). Required,

xpred_trt
A data frame or matrix containing out of sample prognostic covariates. Observations must be in columns. If data frame is set up in usual observations (rows) and covariates (columns) form, input t(x). Required,

tpred
A vector with out of sample discrete time points. Required,

zpred
A vector of 0/1 indicators for each out of sample observation. 1 = treatment group, 0 = control group. Required,

xinfo_list
A list with length matching number of covariates in x. Each element contains a vector of cutpoints. Can be generated using the helper function makeCutPoints(). Required,

xinfo_list_trt
A list with length matching number of covariates in x_trt. Each element contains a vector of cutpoints. Can be generated using the helper function makeCutPoints(). Required,

nburn
Number of burn-in MCMC iterations.

nsim
Number of MCMC iterations to save after burn-in.

ntree
Number of trees. Defaults to 200.

lambda
Double lambda hyperparameter for sigma^2 prior. When setting, use estimated sigma2 from scaled/centered response if using scale_y=TRUE input.

nu
Double nu hyperparameter for sigma^2 prior. Required,

ecross
Double number of expected times f(x,t) crosses the mean response over time, alpha(t). We suggest a default value of 1.

ecross_trt
Double number of expected crossings for treatment effect. Required,

treef_name_
Character vector containing file name/path for tree output. Must be populated, even if save_trees is set to FALSE. (In this case, parameter is ignored.) Required,

save_trees
Boolean for writing trees to a .txt file. Defaults to true. Defaults to FALSE.

silent_mode
Boolean for writing progress report to console. Defaults to FALSE.

Value

A list containing the following items:

mcmcdraws

A matrix containing the in-sample MCMC draws for the tsbart fit. Rows are MCMC iterations, columns are observations.

mcmcdraws_trt

A matrix containing the in-sample MCMC draws for the treatment effect. Rows are MCMC iterations, columns are observations.

mcmcdraws_oos

A matrix containing the out-of-sample MCMC draws for the tsbart fit. Rows are MCMC iterations, columns are observations.

mcmcdraws_oos_trt

A matrix containing the out-of-sample MCMC draws for the treatment effect. Rows are MCMC iterations, columns are observations.

eta_m

A vector containing the MCMC draws for eta_m .

eta_a

A vector containing the MCMC draws for eta_a .

gamma_m

A vector containing the MCMC draws for gamma_m .

gamma_a

A vector containing the MCMC draws for gamma_a .

postpred

A matrix of posterior predictive draws for in-sample data. Rows are MCMC iterations, columns are observations.

postpred_oos

A matrix of posterior predictive draws for out-of-sample data. Rows are MCMC iterations, columns are observations.

Examples

# NOT RUN {
See vignette for full example, including expected crossings parameter tuning.
Function call example is as follows.

Function call example is as follows.

# Calibrate BART's error variance a la CGM 2010 (method 2)
df = data.frame(xx, y)
lmf = lm(y~.,df)
sighat = sigma(lmf) #sd(y) #summary(lmf)$sigma

# Hyperparameters
nu = 3
sigq = .9
qchi = qchisq(1.0-sigq,nu)
lambda = (sighat*sighat*qchi)/nu

# Decision rule cutpoints for each predictor.
cutpoints = makeCutpoints(xx)

# Fit model
fit = tsbcfProbit(y = y,
                    yobs = yobs,
                    x = t(xx),     # observations must be in columns
                    x_trt = t(xx),
                    tgt = ti,
                    z = z,
                    xpred = t(xpred),
                    xpred_trt = t(xpred),
                    tpred = ti_pred,
                    zpred = z_pred,
                    xinfo_list = cutpoints,
                    xinfo_list_trt = cutpoints,
                    nburn = nburn, nsim = nsim, ntree = ntree, m_trt = m_trt,
                    lambda=lambda, nu=nu, ecross = exp_cross, ecross_trt = .5,
                    treef_name_ =  paste0(out.csv,'trees.txt'))

# }