tsbcf: Fits tsbcf model for continuous or binary responses.

This function implements the tsbart method for a continuous response in the causal inference setting, smoothing over the covariate tgt, with unsmoothed covariates x and treatment indicator z. Predicts function values f(x,t,z) at a set of new time points and covariates.

Augmented model formulation is: y(t) = alpha(t) + eta_control * m(x,t) + eta_moderate * a(x,t)*z + epsilon

In the case of binary response, all

tsbcf(y, pihat, pihatpred, z, zpred, tgt, tpred,
         x_control, x_moderate,
         xpred_control, xpred_moderate,
         nburn, nsim, ntree_control=200, ntree_moderate=200,
         lambda=NULL, sigq=.9, sighat=NULL, nu=3,
         base_control=.95, power_control=2,
         base_moderate=.25, power_moderate=3,
         ecross=1, probit=FALSE, yobs=NULL, verbose=T)

Arguments

y
Length n vector with continuous response values.

pihat
Length n vector with propensity score estimates.

pihatpred
Length npred vector with out-of-sample propensity score estimates.

z
Length n vector of 0/1 indicators for each observation. 1 = treatment group, 0 = control group.

zpred
Length npred vector of 0/1 indicators for each out-of-sample observation. 1 = treatment group, 0 = control group.

tgt
Length n vector with discrete time points.

tpred
Length npred vector with out-of-sample discrete time points.

x_control
A (n x p_control) data frame or matrix containing the covariates for the prognostic effect which are not to be smoothed over. Observations must be in columns. If data frame is set up in usual observations (rows) and covariates (columns) form, input t(x).

x_moderate
A (n x p_moderate) data frame or matrix containing the covariates for the treatment effect which are not to be smoothed over. Observations must be in columns. If data frame is set up in usual observations (rows) and covariates (columns) form, input t(x).

xpred_control
A (npred x p_control) data frame or matrix containing out-of-sample covariates. Observations must be in columns. If data frame is set up in usual observations (rows) and covariates (columns) form, input t(x).

xpred_moderate
A (npred x p_moderate) data frame or matrix containing out-of-sample prognostic covariates. Observations must be in columns. If data frame is set up in usual observations (rows) and covariates (columns) form, input t(x).

nburn
Number of burn-in MCMC iterations.

nsim
Number of MCMC iterations to save after burn-in.

ntree_control
Number of trees for control fit.

ntree_moderate
Number of trees for treatment fit.

lambda
Scale parameter in the chisq prior sigma^2. Defaults to NULL, ie being estimated from sigq and sighat. Not appicable for binary case.

sigq
Calibration quantile in the chisq prior on sigma^2. Defaults to .9. Not applicable for binary case.

sighat
Calibration estimate for chisq prior on sigma^2. Defaults to being estimated from data using linear model. Not applicable for binary case.

nu
Degrees of freedom in the chisq prior on sigma^2. Defaults to 3. Not applicable for binary case.

base_control
Base for prognostic tree prior. Defaults to 0.95.

power_control
Power for prognostic tree prior. Defaults to 2.0.

base_moderate
Base for treatment tree prior. Defaults to 0.25.

power_moderate
Power for treatment tree prior. Defaults to 3.0.

ecross
Number of expected times f(x,t) crosses the mean response over time for prognostic fit. We suggest a default value of 1.

probit
0 indicates continuous response; 1 indicates probit response. Default is 0. If 1, then yobs must be populated.

yobs
Length n vector of binary responses; only populated for probit=1 case.

verbose
Boolean for writing progress report to console.

Value

A list containing the following items:

mcmcdraws_control

A (nsim x n) matrix containing the in-sample MCMC draws for the progostic effect.

mcmcdraws_moderate

A (nsim x n) matrix containing the in-sample MCMC draws for the treatment effect.

mcmcdraws_oos_control

A (nsim x npred) matrix containing the out-of-sample MCMC draws for the tsbart fit.

mcmcdraws_oos_moderate

A (nsim x npred) matrix containing the out-of-sample MCMC draws for the prognostic.

sigma

A vector containing the MCMC draws for sigma. Returns 1 in the probit case.

eta_control

A vector containing the MCMC draws for prognostic eta.

eta_moderate

A vector containing the MCMC draws for treatment eta.

gamma_control

A vector containing the MCMC draws for prognostic gamma.

gamma_moderate

A vector containing the MCMC draws for treatment gamma.

postpred

A (nsim x n) matrix of posterior predictive draws for in-sample data.

postpred_oos

A (nsim x npred) matrix of posterior predictive draws for out-of-sample data.

Examples

# NOT RUN {
Function call example is as follows.

# Fit model
fit = tsbcf(y, pihat, pihatpred, z, zpred, tgt, tpred,
                  x_control, x_moderate,
                  xpred_control, xpred_moderate,
                  nburn, nsim, ntree_control=200, ntree_moderate=200,
                  lambda=NULL, sigq=.9, sighat=NULL, nu=3,
                  base_control=.95, power_control=2,
                  base_moderate=.25, power_moderate=3,
                  ecross=1, probit=FALSE, yobs=NULL, verbose=T)

# }